+++ src/GameFinder.Common/IGame.cs
@@ -1,0 +1,9 @@
+using JetBrains.Annotations;
+
+namespace GameFinder.Common;
+
+/// <summary>
+/// Interface for games.
+/// </summary>
+[PublicAPI]
+public interface IGame { }