+++ src/GameFinder.StoreHandlers.EADesktop/EADesktopGame.cs
@@ -1,3 +1,4 @@
+using GameFinder.Common;
 using JetBrains.Annotations;
 using NexusMods.Paths;
 
@@ -10,7 +11,7 @@ namespace GameFinder.StoreHandlers.EADesktop;
 /// <param name="BaseSlug">Slug name of the game.</param>
 /// <param name="BaseInstallPath">Absolute path to the game folder.</param>
 [PublicAPI]
-public record EADesktopGame(EADesktopGameId EADesktopGameId, string BaseSlug, AbsolutePath BaseInstallPath)
+public record EADesktopGame(EADesktopGameId EADesktopGameId, string BaseSlug, AbsolutePath BaseInstallPath) : IGame
 {
     /// <summary>
     /// Returns the absolute path to the <c>installerdata.xml</c> file inside the <c>__Installer</c> folder